Restructure client routing (#157)
* update

* update react-router, delete react-router-redux

* add Layout component

* delete onExplore

* get rid of pathname conditions

* move topnav

* rename App to CLNFetcher
