+++ client/package.json
@@ -43,7 +43,6 @@
     "mini-css-extract-plugin": "^0.4.0",
     "moment": "^2.23.0",
     "node-sass": "^4.9.0",
-    "moment": "^2.23.0",
     "osseus-wallet": "0.0.6",
     "prop-types": "^15.6.1",
     "react": "^16.6.3",
