+++ client/src/containers/Root.js
@@ -5,12 +5,15 @@ import { Route } from 'react-router'
 import { ConnectedRouter } from 'connected-react-router'
 import createHistory from 'history/createBrowserHistory'
 import { AnimatedRoute } from 'react-router-transition'
-import App from 'containers/App'
+import CLNFetcher from 'containers/CLNFetcher'
+import Oven from 'components/oven/Oven'
 import IssuanceWizard from 'components/issuance/IssuanceWizard'
 import ContactForm from 'components/ContactForm'
 import Dashboard from 'components/Dashboard'
 import withTracker from 'containers/withTracker'
 import Web3, {withNetwork} from 'containers/Web3'
+import Layout from 'components/Layout'
+import 'scss/styles.scss'
 
 const history = createHistory()
 
@@ -45,21 +48,25 @@ export default class Root extends Component {
           <div>
             <Web3 />
             <div style={{height: '100%'}}>
-              <Route path='/' component={withTracker(withNetwork(App))} />
-              <div className='contact-form-wrapper'>
-                <AnimatedRoute
-                  path='/view/contact-us'
-                  component={withTracker(ContactForm)}
+              <Layout>
+                <Route path='/' component={withNetwork(CLNFetcher)} />
+                <Route exact path='/' component={withTracker(withNetwork(Oven))} />
+                <div className='contact-form-wrapper'>
+                  <AnimatedRoute
+                    atLeave={0}
+                    path='/view/contact-us'
+                    component={withTracker(ContactForm)}
+                    mapStyles={mapStylesContact}
+                    {...contactFormTransition}
+                  />
+                </div>
+                <Route
+                  path='/view/issuance'
+                  component={withTracker(withNetwork(IssuanceWizard))}
                   mapStyles={mapStylesContact}
                   {...contactFormTransition}
                 />
-              </div>
-              <Route
-                path='/view/issuance'
-                component={withTracker(withNetwork(IssuanceWizard))}
-                mapStyles={mapStylesContact}
-                {...contactFormTransition}
-              />
+              </Layout>
               <Route
                 path='/view/dashboard/:address'
                 component={withTracker(withNetwork(Dashboard))}
