+++ client/src/scss/components/Root.scss
@@ -1,4 +1,5 @@
 .contact-form-wrapper {
+  z-index: 1;
 	border-radius: 2px;
 	position: absolute;
 	//overflow: scroll;
